new try

+++ exercises.c
@@ -29,7 +29,7 @@ void arrayMaxMin(int *a, int n, int *max, int *min) {
   *max = a[0];
   *min = a[0];
   
-  for (size_t i = 1; i < n; i++)
+  for (int i = 1; i < n; i++)
   {
     if (a[i] > *max)
         *max = a[i];
